Beginning framework for equation manager up.
